AutoCoroner now deletes corpses and resets itself
No FX on that yet

Also cleanup any instances in the project where a class did not have a file
