+++ Assets/Prefabs/Obstacles/JailDoor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b6e4922a93e76ff4d977d052fcac63dc
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
