+++ Assets/Scripts/Consequences/TagAffecting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0c3079289ba4ad438f431f593af1a68
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
