+++ Assets/Scripts/Consequences/TagAffecting/DestroyGameObjectsByTagConsequence.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5558ac7f41748544aaf1f40859f7f923
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
